extract the class
